add thew new midddleware of is owner middleware, with also i make sure send a correct status code of 5xx and 4xx group,with also i fix the cron task of background jobs

+++ backend/controllers/saveJobs.controller.js
@@ -14,13 +14,12 @@ const storeSaveJob=async (req, res)=>{
     return res.status(404).json({message: "Please Enter Jobs id to bookmarked jobs"});
   }
   try {
-    
     const {rowCount}=await connect.query("select job_id, users_id from saved_jobs where job_id=$1 and users_id=$2", [id, uid]);
     if(rowCount>0){
-      return res.status(401).json({message: "The Jobs is already Bookmarked don't need to bookmark again"});
+      return res.status(400).json({message: "The Jobs is already Bookmarked don't need to bookmark again"});
     }
-    const {rows}=await connect.query("insert into saved_jobs (job_id, users_id) values ($1, $2) returning *", [id, uid])
-    res.status(201).json({message: rows[0]})
+    await connect.query("insert into saved_jobs (job_id, users_id) values ($1, $2) returning *", [id, uid])
+    return res.status(201).json({message: "Jobs have been bookmarked"})
   } catch (error) {
     return res.status(403).json({message: error.message})
   }
@@ -33,6 +32,10 @@ const unsaveListJob=async (req, res)=>{
     return res.status(404).json({message: "Please Enter Jobs id to bookmarked jobs"});
   }
   try {
+      const {rowCount}=await connect.query("select job_id from saved_jobs where job_id=$1 and users_id=$2", [id, uid]);
+    if(rowCount==0){
+      return res.status(400).json({message: "Please first Bookmark to remove from bookmark"});
+    }
     await connect.query("delete from saved_jobs where job_id=$1 and users_id=$2", [id, uid]);
     return res.status(201).json({message: 'Successfully Removed jobs from bookmark'})
   } catch (error) {
